Refactor: unitDecorator > types 파일 기능별로 분리
