--- src/plugins/unitDecorator/types.ts
@@ -1,134 +1,0 @@
-import { blue } from '@material-ui/core/colors';
-import { deepOrange } from 'themes/colors/deepOrange';
-import { palette } from 'themes/palatte';
-
-export const UNITS = {
-  ROW: '코',
-  STITCH: '단',
-  REPEAT: '번',
-} as const;
-
-export type UNIT_TYPE = typeof UNITS[keyof typeof UNITS];
-
-export const NOT_CALCULATE = 'NOT_CALCULATE';
-
-const UNIT_DISPLAY_APPROXIMATION = {
-  ROUND_DOWN: '내림',
-  ROUND: '반올림',
-  ROUND_UP: '올림',
-  NOT_CALCULATE: '계산 안함',
-} as const;
-
-export const STITCH_APPROXIMATION = {
-  STITCH_CALCULATE_ROUND_DOWN: 'STITCH_CALCULATE_ROUND_DOWN',
-  STITCH_CALCULATE_ROUND: 'STITCH_CALCULATE_ROUND',
-  STITCH_CALCULATE_ROUND_UP: 'STITCH_CALCULATE_ROUND_UP',
-} as const;
-
-export const STITCH_DISPLAY_APPROXIMATION = {
-  [STITCH_APPROXIMATION.STITCH_CALCULATE_ROUND_DOWN]:
-    UNIT_DISPLAY_APPROXIMATION.ROUND_DOWN,
-  [STITCH_APPROXIMATION.STITCH_CALCULATE_ROUND]:
-    UNIT_DISPLAY_APPROXIMATION.ROUND,
-  [STITCH_APPROXIMATION.STITCH_CALCULATE_ROUND_UP]:
-    UNIT_DISPLAY_APPROXIMATION.ROUND_UP,
-  [NOT_CALCULATE]: UNIT_DISPLAY_APPROXIMATION.NOT_CALCULATE,
-} as const;
-
-export type STITCH_APPROXIMATION_TYPE = typeof STITCH_APPROXIMATION[keyof typeof STITCH_APPROXIMATION];
-
-export const ROW_APPROXIMATION = {
-  ROW_CALCULATE_ROUND_DOWN: 'ROW_CALCULATE_ROUND_DOWN',
-  ROW_CALCULATE_ROUND: 'ROW_CALCULATE_ROUND',
-  ROW_CALCULATE_ROUND_UP: 'ROW_CALCULATE_ROUND_UP',
-} as const;
-
-export const ROW_DISPLAY_APPROXIMATION = {
-  [ROW_APPROXIMATION.ROW_CALCULATE_ROUND_DOWN]:
-    UNIT_DISPLAY_APPROXIMATION.ROUND_DOWN,
-  [ROW_APPROXIMATION.ROW_CALCULATE_ROUND]: UNIT_DISPLAY_APPROXIMATION.ROUND,
-  [ROW_APPROXIMATION.ROW_CALCULATE_ROUND_UP]:
-    UNIT_DISPLAY_APPROXIMATION.ROUND_UP,
-  [NOT_CALCULATE]: UNIT_DISPLAY_APPROXIMATION.NOT_CALCULATE,
-} as const;
-
-export type ROW_APPROXIMATION_TYPE = typeof ROW_APPROXIMATION[keyof typeof ROW_APPROXIMATION];
-export type UNIT_APPROXIMATION_TYPE =
-  | STITCH_APPROXIMATION_TYPE
-  | ROW_APPROXIMATION_TYPE
-  | REPEAT_APPROXIMATION_TYPE
-  | typeof NOT_CALCULATE;
-
-export const REPEAT_APPROXIMATION = {
-  STITCH_REPEAT_CALCULATE_ROUND_DOWN: 'STITCH_REPEAT_CALCULATE_ROUND_DOWN',
-  STITCH_REPEAT_CALCULATE_ROUND: 'STITCH_REPEAT_CALCULATE_ROUND',
-  STITCH_REPEAT_CALCULATE_ROUND_UP: 'STITCH_REPEAT_CALCULATE_ROUND_UP',
-  ROW_REPEAT_CALCULATE_ROUND_DOWN: 'ROW_REPEAT_CALCULATE_ROUND_DOWN',
-  ROW_REPEAT_CALCULATE_ROUND: 'ROW_REPEAT_CALCULATE_ROUND',
-  ROW_REPEAT_CALCULATE_ROUND_UP: 'ROW_REPEAT_CALCULATE_ROUND_UP',
-} as const;
-
-export const REPEAT_DISPLAY_APPROXIMATION = {
-  [REPEAT_APPROXIMATION.STITCH_REPEAT_CALCULATE_ROUND_DOWN]: '코 내림',
-  [REPEAT_APPROXIMATION.STITCH_REPEAT_CALCULATE_ROUND]: '코 반올림',
-  [REPEAT_APPROXIMATION.STITCH_REPEAT_CALCULATE_ROUND_UP]: '코 올림',
-  [REPEAT_APPROXIMATION.ROW_REPEAT_CALCULATE_ROUND_DOWN]: '단 내림',
-  [REPEAT_APPROXIMATION.ROW_REPEAT_CALCULATE_ROUND]: '단 반올림',
-  [REPEAT_APPROXIMATION.ROW_REPEAT_CALCULATE_ROUND_UP]: '단 올림',
-  [NOT_CALCULATE]: '계산 안함',
-} as const;
-
-export type REPEAT_APPROXIMATION_TYPE = typeof REPEAT_APPROXIMATION[keyof typeof REPEAT_APPROXIMATION];
-
-const UNIT_COLORS = {
-  STITCH_ROUND_DOWN: {
-    background: deepOrange[300],
-    color: palette.background.paper,
-  },
-  STITCH_ROUND: {
-    background: palette.primary.main,
-    color: palette.background.paper,
-  },
-  STITCH_ROUND_UP: {
-    background: deepOrange[900],
-    color: palette.background.paper,
-  },
-  ROW_ROUND_DOWN: {
-    background: blue[300],
-    color: palette.background.paper,
-  },
-  ROW_ROUND: {
-    background: palette.secondary.main,
-    color: palette.background.paper,
-  },
-  ROW_ROUND_UP: {
-    background: blue[900],
-    color: palette.background.paper,
-  },
-  NOT_CALCULATE: {
-    background: palette.action.disabledBackground,
-    color: palette.text.primary,
-  },
-};
-
-export const UnitDecoratorStyleMap = {
-  [NOT_CALCULATE]: UNIT_COLORS.NOT_CALCULATE,
-  [STITCH_APPROXIMATION.STITCH_CALCULATE_ROUND]: UNIT_COLORS.STITCH_ROUND,
-  [STITCH_APPROXIMATION.STITCH_CALCULATE_ROUND_UP]: UNIT_COLORS.STITCH_ROUND_UP,
-  [STITCH_APPROXIMATION.STITCH_CALCULATE_ROUND_DOWN]:
-    UNIT_COLORS.STITCH_ROUND_DOWN,
-  [ROW_APPROXIMATION.ROW_CALCULATE_ROUND]: UNIT_COLORS.ROW_ROUND,
-  [ROW_APPROXIMATION.ROW_CALCULATE_ROUND_UP]: UNIT_COLORS.ROW_ROUND_UP,
-  [ROW_APPROXIMATION.ROW_CALCULATE_ROUND_DOWN]: UNIT_COLORS.ROW_ROUND_DOWN,
-  [REPEAT_APPROXIMATION.STITCH_REPEAT_CALCULATE_ROUND]:
-    UNIT_COLORS.STITCH_ROUND,
-  [REPEAT_APPROXIMATION.STITCH_REPEAT_CALCULATE_ROUND_UP]:
-    UNIT_COLORS.STITCH_ROUND_UP,
-  [REPEAT_APPROXIMATION.STITCH_REPEAT_CALCULATE_ROUND_DOWN]:
-    UNIT_COLORS.STITCH_ROUND_DOWN,
-  [REPEAT_APPROXIMATION.ROW_REPEAT_CALCULATE_ROUND]: UNIT_COLORS.ROW_ROUND,
-  [REPEAT_APPROXIMATION.ROW_REPEAT_CALCULATE_ROUND_UP]:
-    UNIT_COLORS.ROW_ROUND_UP,
-  [REPEAT_APPROXIMATION.ROW_REPEAT_CALCULATE_ROUND_DOWN]:
-    UNIT_COLORS.ROW_ROUND_DOWN,
-};
