+++ src/plugins/unitDecorator/utils.ts
@@ -1,12 +1,12 @@
 import {
-  STITCH_DISPLAY_APPROXIMATION,
-  ROW_DISPLAY_APPROXIMATION,
-  REPEAT_APPROXIMATION,
-  REPEAT_DISPLAY_APPROXIMATION,
   UNIT_TYPE,
   UNIT_APPROXIMATION_TYPE,
-  ROW_APPROXIMATION,
   STITCH_APPROXIMATION,
+  ROW_APPROXIMATION,
+  REPEAT_APPROXIMATION,
+  STITCH_DISPLAY_APPROXIMATION,
+  ROW_DISPLAY_APPROXIMATION,
+  REPEAT_DISPLAY_APPROXIMATION,
 } from './types';
 
 export const getCalculateKey = (unit?: UNIT_TYPE): string => {
